# backend/docs/swagger.json
{
  "basePath": "/api/v1",
  "definitions": {
...
    "services.LightweightPaymentRequest": {
      "properties": {
        "amount": {
          "type": "number"
        },
        "cash_bank_id": {
          "type": "integer"
        },
        "method": {
          "type": "string"
        },
        "notes": {
          "type": "string"
        },
        "payment_date": {
          "type": "string"
        },
        "reference": {
          "type": "string"
        },
        "sale_id": {
          "type": "integer"
        }
      },
      "required": [
        "amount",
        "cash_bank_id",
        "method",
        "payment_date",
        "sale_id"
      ],
      "type": "object"
    },
    "services.LightweightPaymentResponse": {
      "properties": {
        "amount": {
          "type": "number"
        },
        "message": {
          "type": "string"
        },
        "new_status": {
          "type": "string"
        },
        "outstanding_amount": {
          "type": "number"
        },
        "payment_code": {
          "type": "string"
        },
        "payment_id": {
          "type": "integer"
        },
        "processing_time": {
          "type": "string"
        },
        "success": {
          "type": "boolean"
        }
      },
      "type": "object"
    },
...
  },
  "host": "localhost:8080",
  "info": {
    "contact": {
      "email": "support@swagger.io",
      "name": "API Support",
      "url": "http://www.swagger.io/support"
    },
    "description": "API untuk aplikasi sistem akuntansi yang komprehensif dengan fitur lengkap manajemen keuangan, inventory, sales, purchases, dan reporting.",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "termsOfService": "http://swagger.io/terms/",
    "title": "Sistema Akuntansi API",
    "version": "1.0"
  },
  "paths": {
...
    "/api/v1/payments/fast/record": {
      "post": {
        "consumes": [
          "application/json"
        ],
        "description": "Records a payment with minimal database operations for fast response",
        "parameters": [
          {
            "description": "Payment details",
            "in": "body",
            "name": "payment",
            "required": true,
            "schema": {
              "$ref": "#/definitions/services.LightweightPaymentRequest"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/services.LightweightPaymentResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "additionalProperties": true,
              "type": "object"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "additionalProperties": true,
              "type": "object"
            }
          }
        },
        "summary": "Record payment with fast processing",
        "tags": [
          "fast-payments"
        ]
      }
    },
    "/api/v1/payments/fast/record-async": {
      "post": {
        "consumes": [
          "application/json"
        ],
        "description": "Records payment immediately and creates journal entry in background",
        "parameters": [
          {
            "description": "Payment details",
            "in": "body",
            "name": "payment",
            "required": true,
            "schema": {
              "$ref": "#/definitions/services.LightweightPaymentRequest"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/services.LightweightPaymentResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "additionalProperties": true,
              "type": "object"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "additionalProperties": true,
              "type": "object"
            }
          }
        },
        "summary": "Record payment with async journal processing",
        "tags": [
          "fast-payments"
        ]
      }
    },
    "/api/v1/payments/fast/status/{id}": {
      "get": {
        "description": "Gets real-time payment processing status",
        "parameters": [
          {
            "description": "Payment ID",
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer"
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Payment status",
            "schema": {
              "additionalProperties": true,
              "type": "object"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "additionalProperties": true,
              "type": "object"
            }
          },
          "404": {
            "description": "Payment not found",
            "schema": {
              "additionalProperties": true,
              "type": "object"
            }
          }
        },
        "summary": "Get payment status",
        "tags": [
          "fast-payments"
        ]
      }
    },
    "/api/v1/payments/fast/validate": {
      "post": {
        "consumes": [
          "application/json"
        ],
        "description": "Validates payment data quickly for frontend feedback",
        "parameters": [
          {
            "description": "Payment details",
            "in": "body",
            "name": "payment",
            "required": true,
            "schema": {
              "$ref": "#/definitions/services.LightweightPaymentRequest"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Validation result",
            "schema": {
              "additionalProperties": true,
              "type": "object"
            }
          },
          "400": {
            "description": "Validation errors",
            "schema": {
              "additionalProperties": true,
              "type": "object"
            }
          }
        },
        "summary": "Validate payment data",
        "tags": [
          "fast-payments"
        ]
      }
    },
...
  },
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "BearerAuth": {
      "description": "Type \"Bearer\" followed by a space and JWT token.",
      "in": "header",
      "name": "Authorization",
      "type": "apiKey"
    }
  },
  "swagger": "2.0"
}